+++ src/semantictokens.js
@@ -1,6 +1,7 @@
 "use strict"
 
 const vscode = require('vscode')
+const common = require('./common')
 const selector = { language: 'couper' }
 const { functions, variables } = require('./schema')
 
@@ -9,29 +10,35 @@ function createRegex(list) {
 }
 
 const Regexes = {
-	variable: createRegex(Object.keys(variables)),
+	variable: createRegex(Object.keys(variables).map(name => "(?<!\\.)" + name)),
 	function: createRegex(Object.keys(functions).map(name => name + "\\s*(?=\\()"))
 }
 
 const tokenTypes = Object.keys(Regexes)
+const tokenModifiers = ['defaultLibrary', 'readonly'];
 
-function addTokens(tokensBuilder, document, regex, type) {
+function addTokens(tokensBuilder, document, regex, type, modifiers) {
 	const text = document.getText()
 	const matches = text.matchAll(regex)
 	for (const match of matches) {
 		const start = document.positionAt(match.index)
 		const end = document.positionAt(match.index + match[0].length)
-		tokensBuilder.push(new vscode.Range(start, end), type)
+		const range = new vscode.Range(document.positionAt(0), end)
+		const filteredText = common.filterCommentsAndStrings(document.getText(range))
+		if (filteredText.slice(-match[0].length) == match[0]) {
+			// token is not part of a string or comment
+			tokensBuilder.push(new vscode.Range(start, end), type, modifiers)
+		}
 	}
 }
 
-const legend = new vscode.SemanticTokensLegend(tokenTypes)
+const legend = new vscode.SemanticTokensLegend(tokenTypes, tokenModifiers)
 
 const semanticTokensProvider = {
 	provideDocumentSemanticTokens(document) {
 		const tokensBuilder = new vscode.SemanticTokensBuilder(legend)
-		addTokens(tokensBuilder, document, Regexes.function, 'function')
-		addTokens(tokensBuilder, document, Regexes.variable, 'variable')
+		addTokens(tokensBuilder, document, Regexes.function, 'function', ['defaultLibrary'])
+		addTokens(tokensBuilder, document, Regexes.variable, 'variable', ['readonly'])
 		return tokensBuilder.build()
 	}
 }